requirements.txt
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
cryptography==42.0.2
defusedxml==0.7.1
distlib==0.3.8
dj-rest-auth==5.0.2
Django==5.0.1
django-allauth==0.57.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
filelock==3.13.1
identify==2.5.33
idna==3.6
nodeenv==1.8.0
oauthlib==3.2.2
pillow==10.2.0
platformdirs==4.2.0
pre-commit==3.6.0
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.4
urllib3==2.2.0
virtualenv==20.25.0
